+++ lib/nelmon/__init__.py
@@ -1,2 +1,3 @@
+"""Nelmon - Networklore Monitoring Toolkit."""
 __version__ = "1.3.4"
 __author__ = 'Patrick Ogenstad'
