+++ lib/nelmon/cli/__init__.py
@@ -1,0 +1,1 @@
+"""nelmon.cli."""
