+++ lib/nelmon/snmp/__init__.py
@@ -1,0 +1,1 @@
+"""nelmon.snmp."""
