+++ lib/nelmon/snmp/oids.py
@@ -1,3 +1,4 @@
+"""nelmon.snmp.oids."""
 from nelsnmp.oids import GeneralOids
 from nelsnmp.vendors.cisco.oids import CiscoOids
 from nelsnmp.vendors.synology.oids import SynologyOids
