+++ pylama.ini
@@ -1,0 +1,7 @@
+[pylama]
+linters = mccabe,pep257,pep8,pyflakes
+ignore = D203,
+skip = .tox/*
+
+[pylama:pep8]
+max_line_length = 110
