+++ tests/test_first.py
@@ -1,5 +1,7 @@
+"""Temporary first test file."""
 from nelmon import constants
 
 
 def test_sample():
+    """Basic sanity test."""
     assert constants.OK == 0
